+++ core/src/main/java/org/elasticsearch/action/search/ScrollIdForNode.java
@@ -17,33 +17,22 @@
  * under the License.
  */
 
-package org.elasticsearch.action.search.type;
+package org.elasticsearch.action.search;
 
-/**
- * Base implementation for an async action.
- */
-public class AbstractAsyncAction {
+class ScrollIdForNode {
+    private final String node;
+    private final long scrollId;
 
-    private final long startTime;
-
-    protected AbstractAsyncAction() {
-        this.startTime = System.currentTimeMillis();
+    public ScrollIdForNode(String node, long scrollId) {
+        this.node = node;
+        this.scrollId = scrollId;
     }
 
-    /**
-     * Return the time when the action started.
-     */
-    protected final long startTime() {
-        return startTime;
+    public String getNode() {
+        return node;
     }
 
-    /**
-     * Builds how long it took to execute the search.
-     */
-    protected final long buildTookInMillis() {
-        // protect ourselves against time going backwards
-        // negative values don't make sense and we want to be able to serialize that thing as a vLong
-        return Math.max(1, System.currentTimeMillis() - startTime);
+    public long getScrollId() {
+        return scrollId;
     }
-
 }
