+++ core/src/main/java/org/elasticsearch/action/search/TransportSearchAction.java
@@ -20,10 +20,6 @@
 package org.elasticsearch.action.search;
 
 import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.action.search.type.TransportSearchDfsQueryAndFetchAction;
-import org.elasticsearch.action.search.type.TransportSearchDfsQueryThenFetchAction;
-import org.elasticsearch.action.search.type.TransportSearchQueryAndFetchAction;
-import org.elasticsearch.action.search.type.TransportSearchQueryThenFetchAction;
 import org.elasticsearch.action.support.ActionFilters;
 import org.elasticsearch.action.support.HandledTransportAction;
 import org.elasticsearch.cluster.ClusterService;
@@ -33,13 +29,14 @@ import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.IndexNotFoundException;
 import org.elasticsearch.indices.IndexClosedException;
+import org.elasticsearch.search.action.SearchServiceTransportAction;
+import org.elasticsearch.search.controller.SearchPhaseController;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportService;
 
 import java.util.Map;
 import java.util.Set;
 
-import static org.elasticsearch.action.search.SearchType.DFS_QUERY_THEN_FETCH;
 import static org.elasticsearch.action.search.SearchType.QUERY_AND_FETCH;
 
 /**
@@ -48,25 +45,18 @@ import static org.elasticsearch.action.search.SearchType.QUERY_AND_FETCH;
 public class TransportSearchAction extends HandledTransportAction<SearchRequest, SearchResponse> {
 
     private final ClusterService clusterService;
-    private final TransportSearchDfsQueryThenFetchAction dfsQueryThenFetchAction;
-    private final TransportSearchQueryThenFetchAction queryThenFetchAction;
-    private final TransportSearchDfsQueryAndFetchAction dfsQueryAndFetchAction;
-    private final TransportSearchQueryAndFetchAction queryAndFetchAction;
+    private final SearchServiceTransportAction searchService;
+    private final SearchPhaseController searchPhaseController;
 
     @Inject
-    public TransportSearchAction(Settings settings, ThreadPool threadPool,
-                                 TransportService transportService, ClusterService clusterService,
-                                 TransportSearchDfsQueryThenFetchAction dfsQueryThenFetchAction,
-                                 TransportSearchQueryThenFetchAction queryThenFetchAction,
-                                 TransportSearchDfsQueryAndFetchAction dfsQueryAndFetchAction,
-                                 TransportSearchQueryAndFetchAction queryAndFetchAction,
-                                 ActionFilters actionFilters, IndexNameExpressionResolver indexNameExpressionResolver) {
+    public TransportSearchAction(Settings settings, ThreadPool threadPool, SearchPhaseController searchPhaseController,
+                                 TransportService transportService, SearchServiceTransportAction searchService,
+                                 ClusterService clusterService, ActionFilters actionFilters, IndexNameExpressionResolver
+                                             indexNameExpressionResolver) {
         super(settings, SearchAction.NAME, threadPool, transportService, actionFilters, indexNameExpressionResolver, SearchRequest::new);
+        this.searchPhaseController = searchPhaseController;
+        this.searchService = searchService;
         this.clusterService = clusterService;
-        this.dfsQueryThenFetchAction = dfsQueryThenFetchAction;
-        this.queryThenFetchAction = queryThenFetchAction;
-        this.dfsQueryAndFetchAction = dfsQueryAndFetchAction;
-        this.queryAndFetchAction = queryAndFetchAction;
     }
 
     @Override
@@ -75,7 +65,8 @@ public class TransportSearchAction extends HandledTransportAction<SearchRequest,
         try {
             ClusterState clusterState = clusterService.state();
             String[] concreteIndices = indexNameExpressionResolver.concreteIndices(clusterState, searchRequest);
-            Map<String, Set<String>> routingMap = indexNameExpressionResolver.resolveSearchRouting(clusterState, searchRequest.routing(), searchRequest.indices());
+            Map<String, Set<String>> routingMap = indexNameExpressionResolver.resolveSearchRouting(clusterState,
+                    searchRequest.routing(), searchRequest.indices());
             int shardCount = clusterService.operationRouting().searchShardsCount(clusterState, concreteIndices, routingMap);
             if (shardCount == 1) {
                 // if we only have one group, then we always want Q_A_F, no need for DFS, and no need to do THEN since we hit one shard
@@ -86,16 +77,28 @@ public class TransportSearchAction extends HandledTransportAction<SearchRequest,
         } catch (Exception e) {
             logger.debug("failed to optimize search type, continue as normal", e);
         }
-        if (searchRequest.searchType() == DFS_QUERY_THEN_FETCH) {
-            dfsQueryThenFetchAction.execute(searchRequest, listener);
-        } else if (searchRequest.searchType() == SearchType.QUERY_THEN_FETCH) {
-            queryThenFetchAction.execute(searchRequest, listener);
-        } else if (searchRequest.searchType() == SearchType.DFS_QUERY_AND_FETCH) {
-            dfsQueryAndFetchAction.execute(searchRequest, listener);
-        } else if (searchRequest.searchType() == SearchType.QUERY_AND_FETCH) {
-            queryAndFetchAction.execute(searchRequest, listener);
-        } else {
-            throw new IllegalStateException("Unknown search type: [" + searchRequest.searchType() + "]");
+
+        AbstractSearchAsyncAction searchAsyncAction;
+        switch(searchRequest.searchType()) {
+            case DFS_QUERY_THEN_FETCH:
+                searchAsyncAction = new SearchDfsQueryThenFetchAsyncAction(logger, searchService, clusterService,
+                        indexNameExpressionResolver, searchPhaseController, threadPool, searchRequest, listener);
+                break;
+            case QUERY_THEN_FETCH:
+                searchAsyncAction = new SearchQueryThenFetchAsyncAction(logger, searchService, clusterService,
+                        indexNameExpressionResolver, searchPhaseController, threadPool, searchRequest, listener);
+                break;
+            case DFS_QUERY_AND_FETCH:
+                searchAsyncAction = new SearchDfsQueryAndFetchAsyncAction(logger, searchService, clusterService,
+                        indexNameExpressionResolver, searchPhaseController, threadPool, searchRequest, listener);
+                break;
+            case QUERY_AND_FETCH:
+                searchAsyncAction = new SearchQueryAndFetchAsyncAction(logger, searchService, clusterService,
+                        indexNameExpressionResolver, searchPhaseController, threadPool, searchRequest, listener);
+                break;
+            default:
+                throw new IllegalStateException("Unknown search type: [" + searchRequest.searchType() + "]");
         }
+        searchAsyncAction.start();
     }
 }
